requirements.txt
docker-py
PyYAML
pep8
pylint
pytest
